requirements.txt
# =============================================================================
# GuidelineGuard — Python Dependencies
# =============================================================================

# -- Web Framework --
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# -- Database --
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# -- AI / LLM --
openai==1.59.5

# -- Vector Search & Embeddings --
faiss-cpu==1.9.0.post1
numpy==1.26.4
torch==2.2.2
transformers==4.47.1

# -- Data Processing --
pandas==2.2.3

# -- HTTP --
httpx==0.28.1

# -- Testing --
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# -- CORS --
# (included in fastapi)
